Fix the IMU timestamp bug
Change the method used in the camera callbacks in order to get the
nanoseconds since the OS booted for consistency.

According to the Android's API Reference, the SensorEvent class holds
information including timestamp at which the event happened in
nanosecond. But it is actually the number of nanoseconds since the
operating system started.